// src/interfaces/pokemonInterfaces.tsx
export interface SimplePokemon {
    id: string;
    name: string;
    color?: string;

    picture:  string;
}